Commiting changes; hard drive is dying :O

+++ src/common/utils.rs
@@ -6,15 +6,15 @@ pub fn hash<T: Hash>(t: &T) -> u64 {
     s.finish()
 }
 
-pub fn bit_clr(x:&mut u32, y:u8) {
+pub fn bit_clr(x:&mut u32, y: u32) {
     *x = *x & !(1<<y)
 }
 
-pub fn bit_set(x:&mut u32, y:u8) {
+pub fn bit_set(x:&mut u32, y: u32) {
     *x = *x | 1<<y
 }
 
-pub fn is_bit_set(x:u32, y:u8) -> bool {
+pub fn is_bit_set(x: u32, y: u32) -> bool {
     ((x >> y) & 1) == 1
 }
